pyproject.toml
[tool.poetry]
name = "whiskey"
version = "0.1.0"
description = ""
authors = ["deadaf <askpygod@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = {git = "https://github.com/Rapptz/discord.py.git"}
jishaku = {git = "https://github.com/Gorialis/jishaku.git"}
tortoise-orm = "^0.17.6"
async-property = "^0.2.1"
asyncpg = "^0.24.0"
PyNaCl = "^1.4.0"
aiocache = "^0.11.1"
ujson = "^4.0.2"

[tool.poetry.dev-dependencies]
black = {version = "^21.7b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 122